+++ src/components/InstallationPage/components/IntegrationScheme/IntegrationScheme.scss
@@ -210,7 +210,7 @@
       }
 
       &-active {
-        background-color: var(--bg-light-green);
+        background-color: #DEF5CB;
       }
 
       &-first-node {
@@ -239,7 +239,7 @@
         right: -9px;
         border-end-start-radius: 12px;
         border-end-end-radius: 12px;
-        background: var(--bg-white);;
+        background: var(--white);;
         opacity: 1;
         transition: opacity .4s ease-in-out;
         transition-delay: .4s;
@@ -252,11 +252,11 @@
 
       display: block;
       height: auto;
-      border: 1px solid var(--bg-dark-grey);
+      border: 1px solid var(--text-light-grey);
       border-radius: 16px;
       line-height: 0;
       padding: 0 10px;
-      background-color: var(--bg-white);
+      background-color: var(--white);
       color: var(--text-grey);
       
       &-info {
@@ -315,7 +315,7 @@
       right: -5px;
       top: 50%;
       transform: translate(0,-50%);
-      background: var(--bg-white);
+      background: var(--white);
       z-index: 1000px;
     }
   }
@@ -327,7 +327,7 @@
     top: 50%;
     left: -8px;
     right: -8px;
-    background: var(--bg-dark-grey);
+    background: var(--text-light-grey);
     transform: translate(0,-50%);
     z-index: -10;
 
@@ -336,13 +336,13 @@
       width: 5px;
       height: 5px;
       border-radius: 100%;
-      border: 1px solid var(--bg-dark-grey);
+      border: 1px solid var(--text-light-grey);
       display: block;
       position: absolute;
       left: -6px;
       top: 50%;
       transform: translate(0,-50%);
-      background: var(--bg-white);
+      background: var(--white);
       z-index: 1000px;
     }
 
@@ -351,7 +351,7 @@
       display: inline-block;
       border: 5px solid transparent;
       border: 5px solid transparent;
-      border-left: 5px solid var(--bg-dark-grey);
+      border-left: 5px solid var(--text-light-grey);
       top: 50%;
       right: -6px;
       position: absolute;
@@ -364,8 +364,8 @@
     &::before{
       content: "";
       height: 25px;
-      background-color: var(--bg-dark-grey);
-      border-right: 1px solid var(--bg-dark-grey);
+      background-color: var(--text-light-grey);
+      border-right: 1px solid var(--text-light-grey);
       display: block;
       left: 50%;
       margin-left: 0.5px;
@@ -378,7 +378,7 @@
       content: '';
       display: inline-block;
       border: 5px solid transparent;
-      border-top: 5px solid var(--bg-dark-grey);
+      border-top: 5px solid var(--text-light-grey);
       bottom: -40px;
       left: 50%;
       margin-left: 1px;
@@ -508,7 +508,7 @@
     top: 50%;
     left: 0;
     right: 0;
-    border-top: 1px solid var(--bg-dark-grey);
+    border-top: 1px solid var(--text-light-grey);
     transform: translate(0,-50%);
     z-index: -10;
   }
@@ -519,11 +519,11 @@
 
     display: flex;
     justify-content: center;
-    border: 1px solid var(--bg-dark-grey);
+    border: 1px solid var(--text-light-grey);
     border-radius: 16px;
     line-height: 0;
     padding: 0 10px;
-    background-color: var(--bg-white);
+    background-color: var(--white);
     color: var(--color-primary-600);
   }
 }
