+++ src/styles/variables.scss
@@ -34,9 +34,6 @@ $desktop-lg: 1919px;
   --bg-sea-wave: #ccf2f4;
   --bg-grey2: #f4f5fa;
   --bg-light-grey: #fbfcfd;
-  --bg-dark-grey: #c3c8d5;
-  --bg-white: var(--white);
-  --bg-light-green: #DEF5CB;
 
   --graphics-amber: #ffc800;
   --graphics-orange: #ff9900;
